Added full stops and removed FIXME.
Since some of the arguments have default values
we can't reliably check that only full_conf was provided
(w/o other values) - unless we resort to **kwargs signature.
Leave it as it is for now - there is a good chance full_conf
will be removed in the future since currently I don't see
much value in it (and it's not used in my projects).